% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agents_required.R
\name{agents_required}
\alias{agents_required}
\title{Resource planning for queues}
\usage{
agents_required(demandqty, perxmin = 30, serviceduration, targetservetime,
  targetservicelevel)
}
\arguments{
\item{demandqty}{a numeric vector of the number of demand units entering the queue per time interval}

\item{perxmin}{an integer value indicating the number of minutes the demand qty is measured over.}

\item{serviceduration}{a numeric vector of the average service duration in seconds}

\item{targetservetime}{a numeric vector of the targeted servce time in seconds}

\item{targetservicelevel}{a floating point decimal of the target Grade of Service percentage i.e. percent of traffic within targetserve duration}
}
\value{
a dataframe containing Agents_Required and attainable Service_Level
}
\description{
This algorithm calculates the number of agents required to staff a queue
to achieve a given service level using the Erlang C traffic formula over an M/M/c queue model.
}
\details{
The function accepts the specified service levels and demand statistics and calculates the
probability of queueing using the Erlang C formula. The algorithm iterates the number
of agents present until the target GOS is met.
The model assumes a M/M/c queue system where arrivals form a single queue and are governed by a Poisson
process where there are c servers and service times are exponentially distributed.
}
\author{
Dean Marchiori
}
\references{
http://en.wikipedia.org/wiki/Erlang_unit
}

