% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probQ.R
\name{probQ}
\alias{probQ}
\title{Probability of having to queue in a system}
\usage{
probQ(demandqty, perxmin = 30, serviceduration, numagents)
}
\arguments{
\item{demandqty}{a numeric vector of the number of demand units entering the queue per given time interval}

\item{perxmin}{an integer value indicating the number of minutes the demand qty is measured over.}

\item{serviceduration}{a numeric vector of the average service duration in seconds}

\item{numagents}{an integer value of the number of agents staffing the queue system}
}
\value{
a dataframe containing Agents_In_Use and probability to queue Prob_Queue
}
\description{
This algorithm calculates the probability of having to queue using the Erlang C traffic formula over an M/M/c queue model.
}
\details{
The function accepts the specified service levels and demand statistics and calculates the
probability of queueing using the Erlang C formula.
}
\author{
Dean Marchiori
}
\references{
http://en.wikipedia.org/wiki/Erlang_unit
}

